consumers with autodelivery no longer run out of link credit
A consumer with autoConfirm=true would run out of link credit as
no DeliverResponse would be generated when the DeliveryRequest
was not present in the pendingDeliveries Set.

All deliveries will now generate a DeliveryResponse which will
trigger flow responses back to the broker to ensure the link
credit is not exhausted.