Add SendOptions.retainLink to allow SenderLink to be closed after use
The SendOptions.retainLink is now used to determine if the SenderLink
is retained after sending the outbound message.  By default the Link
will be retained to avoid re-creating the Link on subsequent messages
to the same destination.  Setting the value to false is useful for
'one-off' topics to allow the Link to be closed and resources released.

Fixes #18