Use REPLACED transition for NOT_PERMITTED close
When the server closes a connection with a NotPermittedException, use the
REPLACED transition rather than the STOPPED transition. This is because the
REPLACED transition is used for the server having closed the connection, where
as STOPPED assumes the client has sent a stop and waits for a response from
the server (which in this case we're not going to get).