Add a "fatal" transition into retrying state
Update the FSM to allow a "fatal" response to an open (i.e. server rejects the
client because it is not authorized) when the client is in retrying state. This
is an omission in the original design of the FSM - as we only considered
handling this kind of failure when the client first connects, not when it has
already been connected and is in retrying state.

This fixes the following exception:
```
java.lang.IllegalStateException: No valid leaving transitions are permitted from
state 'Retrying2B' for trigger 'OPEN_RESP_FATAL'. Consider ignoring the trigger.
```